requirements.txt
streamlit
pandas
numpy
langchain
llama-index
llama_index.core
llama-index-agent-openai
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llamaindex-py-client